четвёртое задание
